third_party/WebKit/LayoutTests/editing/execCommand/nsresponder-outdent.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<script>
test(() => assert_selection(
  [
    '<div contenteditable>',
      '|<blockquote class="abc">foo</blockquote>',
    '</div>',
  ].join(''),
  'outdent',
  '<div contenteditable>|foo</div>'),
  'Outdent command with BLOCKQUOTE');
</script>
